BIC-218 # move route handling to 1-step pipeline
